simplify pthread error handling
